@@ -1,15 +1,17 @@
 import CollectionItem from 'Components/collection-item/collection-item.compoent';
 import React from 'react';
-import { useSelector } from 'react-redux';
+import { useAppSelector } from 'Stores/hook';
 import { useParams } from 'react-router';
 import { Collection, selectCollection } from 'Stores/shop/shop.selector';
-import { IShopCollection } from 'Stores/shop/shop1.data';
 import './collection.styles.scss';
+import { IShopCollection } from 'Stores/shop/shop1.data';
 
 const CollectionPage: React.FC = () => {
-	const params: { collection: Collection } = useParams();
-	const collection: IShopCollection = useSelector(selectCollection(params.collection));
-	const { title, items } = collection;
+	const { collection } = useParams();
+	const selectedCollection: IShopCollection = useAppSelector(
+		selectCollection(collection as Collection)
+	);
+	const { title, items } = selectedCollection;
 
 	return (
 		<div className="collection-page">
