@@ -1,12 +1,11 @@
 /* eslint-disable no-console */
-import { RouteComponentProps } from 'react-router';
+import { NavigateFunction } from 'react-router-dom';
 import { FormData } from 'Components/signup/signup.component';
-import { Dispatch } from 'redux';
+import { Dispatch } from '@reduxjs/toolkit';
 import * as api from '../../Api/index';
 
-export const signin =
-	(formData: { email: string; password: string }) =>
-	async (dispatch: Dispatch): Promise<void> => {
+export const signin = (formData: { email: string; password: string }) => {
+	return async (dispatch: Dispatch): Promise<void> => {
 		try {
 			const { data } = await api.signin(formData);
 			dispatch({ type: 'AUTH', data });
@@ -15,14 +14,15 @@ export const signin =
 			console.log(error);
 		}
 	};
+};
 export const signup =
-	(formData: FormData, history: RouteComponentProps['history']) =>
+	(formData: FormData, navigate: NavigateFunction) =>
 	async (dispatch: Dispatch): Promise<void> => {
 		try {
 			const { data } = await api.signup(formData);
 			dispatch({ type: 'AUTH', data });
 			dispatch({ type: 'SET_CURRENT_USER', payload: data });
-			history.push('/');
+			navigate('/');
 		} catch (error) {
 			console.log(error);
 		}
