@@ -1,17 +1,17 @@
-import { combineReducers } from 'redux';
-import storage from 'redux-persist/lib/storage';
-import persistReducer from 'redux-persist/es/persistReducer';
+import { combineReducers } from '@reduxjs/toolkit';
+// import storage from 'redux-persist/lib/storage';
+// import persistReducer from 'redux-persist/es/persistReducer';
 import { authReducer } from './auth/Auth.reducer';
 import { userReducer } from './user/user.reducer';
 import { cartReducer } from './cart/cart.reducer';
 import { directoryReducer } from './directory/directory.reducer';
 import { shopReducer } from './shop/shop.reducer';
 
-const persistConfig = {
-	key: 'root',
-	storage,
-	whitelist: ['cart'],
-};
+// const persistConfig = {
+// 	key: 'root',
+// 	storage,
+// 	whitelist: ['cart'],
+// };
 const rootReducer = combineReducers({
 	user: userReducer,
 	auth: authReducer,
@@ -19,4 +19,5 @@ const rootReducer = combineReducers({
 	directory: directoryReducer,
 	shop: shopReducer,
 });
-export default persistReducer(persistConfig, rootReducer);
+// export default persistReducer(persistConfig, rootReducer);
+export default rootReducer;
