@@ -3,7 +3,7 @@ import { loadStripe } from '@stripe/stripe-js';
 import { Elements } from '@stripe/react-stripe-js';
 
 const stripe = loadStripe(process.env.REACT_APP_STRIPE_PUBLIC_KEY || '');
-const StripeProvider: React.FC = ({ children }) => {
+const StripeProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
 	return <Elements stripe={stripe}>{children}</Elements>;
 };
 
