@@ -1,11 +1,11 @@
 import React, { FC } from 'react';
 import './directory.styles.scss';
-import { useSelector } from 'react-redux';
+import { useAppSelector } from 'Stores/hook';
 import { selectDirectorySections } from 'Stores/directory/directory.selector';
 import MenuItem from '../menu-item/menu-item.component';
 
 const Directory: FC = () => {
-	const sections = useSelector(selectDirectorySections);
+	const sections = useAppSelector(selectDirectorySections);
 	return (
 		<div className="directory-menu">
 			{sections.map(({ id, ...otherSectionProps }) => (
